Выделяем под управление сотрудниками отдельный контроллер
