+++ Controllers/HomeController.cs
@@ -8,52 +8,8 @@ using WebStore.Models;
 namespace WebStore.Controllers
 {    
     public class HomeController : Controller
-    {
-        private static readonly List<Employee> __Employees = new List<Employee>
-        {
-            new Employee
-            {
-                Id = 1,
-                Surname = "Иванов",
-                FirstName = "Иван",
-                Patronymic = "Иванович",
-                Age = 50
-            },
-            new Employee
-            {
-                Id = 2,
-                Surname = "Петров",
-                FirstName = "Петр",
-                Patronymic = "Петрович",
-                Age = 25
-            },
-            new Employee
-            {
-                Id = 3,
-                Surname = "Сидоров",
-                FirstName = "Сидор",
-                Patronymic = "Сидорович",
-                Age = 30
-            }
-        };
-        
-
+    {                
         public IActionResult Index() => View();
-        
-
-        public IActionResult EmployeeDetails(int id)
-        {
-            var employee = __Employees.FirstOrDefault(e => e.Id == id);
-            if (employee is null)
-                return NotFound();
-
-            return View(employee);
-        }
-
-
-        public IActionResult AnotherAction()
-        {
-            return Content("Another action result");
-        }
+                
     }
 }
