+++ Views/Employees/EmployeeDetails.cshtml
@@ -9,5 +9,5 @@
         <p>Возраст: @Model.Age</p>
 
         <hr />
-        <a href="/Home/Index">К списку сотрудников</a>
+        <a asp-controller="Employees" asp-action="Index">К списку сотрудников</a>
     </div>
