+++ Views/Employees/Index.cshtml
@@ -1,0 +1,35 @@
+﻿@model IEnumerable<Employee> 
+
+@{ 
+    ViewData["Title"] = "Index";
+}
+
+<div class="container">
+    <table class="table">
+        <thead>
+            <tr>
+                <th>id</th>
+                <th>Фамилия</th>
+                <th>Имя</th>
+                <th>Отчество</th>
+                <th>Возраст</th>
+                <th></th>
+            </tr>
+        </thead>
+        <tbody>
+            @foreach(var employee in Model)
+            {
+            <tr>
+                <td>@employee.Id</td>
+                <td>@employee.Surname</td>
+                <td>@employee.FirstName</td>
+                <td>@employee.Patronymic</td>
+                <td>@employee.Age</td>
+                @*<td><a href="@Url.Action("EmpoloyeeDetails","Employees")">Информация</a></td>*@
+                <td><a asp-controller="Employees" asp-action="EmployeeDetails" asp-route-id="@employee.Id">Информация</a></td>
+            </tr>
+            }
+        </tbody>
+    </table>
+
+</div>
